software/python/sibila-python/requirements.txt
fastapi
uvicorn[standard]
fastapi-utils
ftfy
cyhunspell
spacy
